Use frames rather than fancy middleware
Most this app's code was middleware - it's much easier just
to use frames. This also avoids the overhead of having
djohno in your production settings, which means it can
safely just be left around, since it won't do anything
unless you visit the djohno URLs, which are locked down to
superusers only anwyay.